fix(mcore): Pass torch_dtype to AutoBridge.from_hf_pretrained for mbridge
When using use_mbridge=True with vanilla_mbridge=False, the model was
being loaded with default dtype (bfloat16) instead of the configured
dtype, causing dtype mismatch errors for ref_module.

This fix passes torch_dtype to AutoBridge.from_hf_pretrained() so the
HuggingFace model is loaded with the correct precision from the start.

This replaces the previous ineffective fix that tried to set
tf_config.fp16/bf16 after model creation.

+++ verl/workers/megatron_workers.py
@@ -178,7 +178,7 @@ class MegatronWorker(Worker):
                 from verl.models.mcore.bridge import AutoBridge
 
                 # Use Megatron-Bridge to convert HF config to Megatron config
-                bridge = AutoBridge.from_hf_pretrained(self.local_path, trust_remote_code=trust_remote_code)
+                bridge = AutoBridge.from_hf_pretrained(self.local_path, trust_remote_code=trust_remote_code, torch_dtype=dtype)
                 # Get Megatron provider and configure it
                 provider = bridge.to_megatron_provider(load_weights=False)
